issues-120 | refactor API module into self-contained structure Move controllers, services, and routes from app/Http/Controllers and app/Services into app/Modules/Api with its own ServiceProvider. Register ApiServiceProvider in bootstrap/providers.php and update all namespace references across the codebase and tests.
------------------------------
Files:
Changed:  app/Console/Commands/GenerateSwagger.php
Changed:  app/Helpers/TelegramHelper.php
Added:    app/Modules/Api/ApiServiceProvider.php
Renamed: app/Http/Controllers/FilesController.php → app/Modules/Api/Controllers/FilesController.php
Renamed: app/Http/Controllers/PreviewController.php → app/Modules/Api/Controllers/PreviewController.php
Renamed: app/Http/Controllers/SimplePage.php → app/Modules/Api/Controllers/SimplePage.php
Renamed: app/Http/Controllers/SwaggerController.php → app/Modules/Api/Controllers/SwaggerController.php
Renamed: app/Services/File/FileService.php → app/Modules/Api/Services/FileService.php
Renamed: app/Services/Swagger/SwaggerGenerateService.php → app/Modules/Api/Services/SwaggerGenerateService.php
Added:    app/Modules/Api/api_routes.php
Added:    app/Modules/Api/web_routes.php
Changed:  bootstrap/providers.php
Changed:  phpstan.neon
Changed:  routes/api.php
Changed:  routes/web.php
Renamed: tests/Unit/Services/File/FileServiceTest.php → tests/Unit/Modules/Api/Services/FileServiceTest.php
