+++ app/Modules/Api/Controllers/PreviewController.php
@@ -1,11 +1,14 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Modules\Api\Controllers;
 
 use Illuminate\Routing\Controller;
 
 class PreviewController extends Controller
 {
+    /**
+     * @return mixed
+     */
     public function chat(): mixed
     {
         return view('preview.chat');
