+++ phpstan.neon
@@ -18,7 +18,7 @@ parameters:
         - '#(?i)referenced with incorrect case#'
         - '#Call to an undefined method Mockery\\ExpectationInterface\|Mockery\\HigherOrderMessage::with\(\)#'
         - identifier: argument.type
-          path: app/Http/Controllers/PreviewController.php
+          path: app/Modules/Api/Controllers/PreviewController.php
 
     reportUnmatchedIgnoredErrors: false
