+++ routes/web.php
@@ -1,19 +1,4 @@
 <?php
 
-use App\Http\Controllers\PreviewController;
-use App\Http\Controllers\SimplePage;
-use App\Http\Controllers\SwaggerController;
-use Illuminate\Support\Facades\Route;
-
-Route::group([
-    'prefix' => 'docs',
-    'as' => 'docs.',
-], function () {
-    Route::get('/swagger-v1-json', [SwaggerController::class, 'showSwagger']);
-    Route::get('/swagger-v1-ui', [SwaggerController::class, 'swaggerUi']);
-});
-
-Route::get('/preview/chat', [PreviewController::class, 'chat']);
-
-Route::get('/', [SimplePage::class, 'index']);
-Route::get('/live_chat_promo', [SimplePage::class, 'liveChatPromo']);
+// Routes are registered by module ServiceProviders:
+// - App\Modules\Api\ApiServiceProvider  → /, /docs/*, /preview/*, /live_chat_promo
